* midpoint double * replaced atomic with mtx due to issues at macos, can be investigated later * const correctness and other minor improvements

+++ include/orderbook/levelOne.hpp
@@ -8,7 +8,7 @@ struct levelOne {
   uint64_t highestBidSize;
   uint64_t lowestAsk;
   uint64_t lowestAskSize;
-  uint64_t midPoint;
+  double midPoint;
   double spreadBps;
 
   bool valid() const {
